add authentication based on SSH agent
This change adds a third authentication method, from ssh-agent to the list
of authentication options.
It will try to retrieve the list of signers based on the SSH_AUTH_SOCK
environment variable and setup an authentication method based on the first
encountered signer in that list.

Also it adds the option to let the tunnel guess the authentication method
automatically. In that case it will try to retrieve a method from a key file
and if that fails through the ssh-agent.